+++ PersistentDataEditor/Controls/StringRender.cs
@@ -9,9 +9,9 @@ internal class StringRender(string showString) : BaseControlItem
 {
     private static Font _font;
     private static Font Font => _font ??= new Font(GH_FontServer.StandardAdjusted.FontFamily, 6);
-    internal override int MinWidth => GH_FontServer.StringWidth(showString, Font);
+    internal override float MinWidth => GH_FontServer.StringWidth(showString, Font);
 
-    internal override int Height => 17;
+    internal override float Height => 17;
 
     internal override void Clicked(GH_Canvas sender, GH_CanvasMouseEvent e)
     {
